chat-terminal-app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
httpx==0.27.0
python-multipart==0.0.9
typing-extensions>=4.8.0
colorama==0.4.6
PyJWT==2.8.0


